+++ tests/EventSourcing.UnitTests/EventRecordTests/Fakes/FakeDomainEvent.cs
@@ -1,0 +1,5 @@
+﻿using ShopSharp.Core.Domain.Events;
+
+namespace ShopSharp.Users.EventSourcing.EventRecordTests.Fakes;
+
+public record FakeDomainEvent : DomainEvent;